build: Enable -Wextra and fix the warnings it surfaces
Loop counters that compare against vec_size() (a size_t) were declared
int, producing -Wsign-compare; make them size_t (library, tests, and
examples). Cast the int expected-values in mtoka_osci's ASSERT_EQ size
checks for the same reason.

Enable -Wextra in the test and example builds, suppressing two classes
that are noise here: -Wno-unused-parameter (unavoidable with the
function-pointer vtable interface) and -Wno-missing-field-initializers
(idiomatic partial designated initializers in test case tables). The
build is now warning-clean under -Wextra.

+++ src/rhc_phase_portrait_plotter.c
@@ -270,7 +270,7 @@ bool ppp_simulator_is_stable(ppp_t *self, vec_t p)
 
 bool __is_lower_any(vec_t v1, vec_t v2)
 {
-  register int i;
+  register size_t i;
 
   for( i=0; i<vec_size(v1); i++ )
     if( vec_elem(v1,i) < vec_elem(v2,i) ) return true;
@@ -279,7 +279,7 @@ bool __is_lower_any(vec_t v1, vec_t v2)
 
 bool __is_greater_any(vec_t v1, vec_t v2)
 {
-  register int i;
+  register size_t i;
 
   for( i=0; i<vec_size(v1); i++ )
     if( vec_elem(v1,i) > vec_elem(v2,i) ) return true;
