+++ test/rhc_mtoka_osci_test.c
@@ -212,21 +212,21 @@ TEST(test_mtoka_osci_init_specify_n_neuron)
   ASSERT_PTRNE( NULL, osci.neurons );
   ASSERT_EQ( 0.0, mtoka_osci_time(&osci_n) );
   ASSERT_EQ( 0, mtoka_osci_step(&osci_n) );
-  ASSERT_EQ( n, vec_size(mtoka_osci_membrane_potential(&osci_n)) );
-  ASSERT_EQ( n, vec_size(mtoka_osci_firing_rate(&osci_n)) );
-  ASSERT_EQ( n, vec_size(mtoka_osci_adapt_property(&osci_n)) );
-  ASSERT_EQ( n, vec_size(mtoka_osci_tonic_input(&osci_n)) );
-  ASSERT_EQ( n, vec_size(mtoka_osci_sensory_feedback(&osci_n)) );
-  ASSERT_EQ( 2*n, vec_size(osci_n.xv) );
+  ASSERT_EQ( (size_t)n, vec_size(mtoka_osci_membrane_potential(&osci_n)) );
+  ASSERT_EQ( (size_t)n, vec_size(mtoka_osci_firing_rate(&osci_n)) );
+  ASSERT_EQ( (size_t)n, vec_size(mtoka_osci_adapt_property(&osci_n)) );
+  ASSERT_EQ( (size_t)n, vec_size(mtoka_osci_tonic_input(&osci_n)) );
+  ASSERT_EQ( (size_t)n, vec_size(mtoka_osci_sensory_feedback(&osci_n)) );
+  ASSERT_EQ( (size_t)( 2*n ), vec_size(osci_n.xv) );
   ASSERT_PTREQ( mtoka_osci_dp, osci_n.ode.f );
   ASSERT_PTRNE( NULL, osci_n.ode._ws );
-  ASSERT_EQ( 2*n, vec_size(((_ode_rk4 *)osci_n.ode._ws)->x) );
+  ASSERT_EQ( (size_t)( 2*n ), vec_size(((_ode_rk4 *)osci_n.ode._ws)->x) );
 
   for( i=0; i<2; i++ ){
     np = mtoka_osci_neuron( &osci_n, i );
     ASSERT_EQ( 1.0, mtoka_osci_neuron_rise_time_const(np) );
     ASSERT_EQ( 1.0, mtoka_osci_neuron_adapt_time_const(np) );
-    ASSERT_EQ( n, vec_size(mtoka_osci_neuron_mutual_inhibit_weights(np)) );
+    ASSERT_EQ( (size_t)n, vec_size(mtoka_osci_neuron_mutual_inhibit_weights(np)) );
     ASSERT_EQ( 0.0, vec_elem(mtoka_osci_neuron_mutual_inhibit_weights(np), 0) );
     ASSERT_EQ( 0.0, vec_elem(mtoka_osci_neuron_mutual_inhibit_weights(np), 1) );
     ASSERT_EQ( 0.0, mtoka_osci_neuron_steady_firing_rate(np) );
